Move launcher token in `tokens.ts` file
